LONG type support

+++ tests/simple.c
@@ -1,8 +1,8 @@
 int main()
 {
-    int a = 2;
-    int b = 5;
-    int c;
+    long a = 2;
+    long b = 5;
+    long c;
 
     c = (a + b) * 2;
     print("Hello world!");
